third_party/blink/web_tests/platform/generic/fast/forms/focus-with-display-block-expected.txt

This test ensures that <output>, <meter> and <progress> are not focused.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

- Moves the focus by using keyDown() in DRT, otherwise using element.focus().
- checkFocus() returns true when <output>, <meter> and <progress> do not have focus.
PASS checkFocus() is true
PASS checkFocus() is true
PASS checkFocus() is true
PASS checkFocus() is true
PASS successfullyParsed is true

TEST COMPLETE

Text in output element
